FEAT-#2197: move concat implementation into general.py
